Change project framework version to 4.0
The build of cbe84bf88 failed in Jenkins with the following error

warning MSB3245: Could not resolve this reference. Could not locate the
assembly "System.Net.Http.WebRequest". Check to make sure the assembly
exists on disk. If this reference is required by your code, you may get
compilation errors.

I opened the project, changed the framework to 4.5, changed it back to 4.0,
and commited the changes.