Display map and pass wrapped map object to plugins
We need to create plugin objects before rendering (so we can render their toolbar names).
But we need to pass a map object to the plugin constructors, and that isn't available until after rendering. (The map object needs to be fully loaded before it knows e.g. its extent.)

Here's what happens:
1) The pane view's render() function constructs an esri.Map object. (It needs a unique DOM id so we create one, since each pane will have its own map.)
2) We then wait until the map is loaded, and tell the pane model to initialize the plugins
3) We use _.extend() to wrap the map object, and pass that to the plugins. (TODO: add override functions in the wrapper)