Fix mistake in Pane.js refactoring
Having an outer "model" or "view" variable doesn't work because it's a single static variable. When there are multiple instances of the class they will interfere with each other. The solution is to pass "model" or "view" to every function. Still perhaps better than the original, but not as clean as before.