Attach the display container when it is added to the model
The rendering of plugin views and the instantiation of the plugin object,
which needs a refrence the dom element to render to, are no syncronously
related.  The quick work around is to let the view attach the rendering div
when it is added to the model.  The el reference on the model is weird, but
I think will be necessary to enforce plugin visibility throug non-UI driven
events.