Extract side bar view and add options for split view
I needed to re-render the sidebar without re-rendering the whole pane (i.e., the map),
so I put sidebar in it's own template.  I added codepaths for proper icon display
when in various main/splitView modes for the sidebar.

Pane 0 or 1 could be main depending on the view mode, and the main pane gets the tool
icons.  Also, when in split mode, I don't show split tool, but rather both Full Map tools.