Validate region.json using a formal JSON schema
Geosite deployers have to edit region.json, so we want to validate it and give good error messages. The JSON reader we're using (Json.Net) includes support for validation using a JSON schema (http://json-schema.org). So I created a schema specifying the structure of region.json (property names, data types, what's required) and used it to validate region.json. The error messages are good, and they're written to the log file.