Add Export/Print feature
create backbone model and view for the Export tool. The model includes
methods to manage calls to the esri api class PrintTask. The model and
view are instantiated from a click event function on the screen view,
and then rendered into a TinyBox modal window.

Modified sidebar template to only show a print button if there is a
corresponding server url in the region.json config. Add click event to
print button to launch the export tool.