Permalink changes
Stop showing permalink in location bar of browser. Passed callbacks
into HashModels in order to override default behavior.

Moved permalink pop-up to App.js to resolve a circular dependency
with storing it in the screen view: the screen view requires hashmodels
to have been initialized, but we want to send the permalink pop-up
callback that was living in the screen view to the init method on
hashmodels.

Highlight permalink text in pop-up box so that users can easily
press their keyboard shortcut for copying.