Add support for plugin getState and setState methods
Screen objects now register their pane models with HashModels
after instantiating them. This takes care of setting the stateOfPlugins
attribute on the Pane, which is unpacked during plugin initialization
state on the plugins. makePermalink is called, the Screen's Panes are
instructed to save their state before HashModels generates a stateString.

Panes call getState and setState on their Plugin wrapper models, which
call getState and setState on the individual plugins themselves.