Wire Layer Selector state into HashModels
Create a currentState object at the plugin level that keeps track
of the visible map layers, opacity, and (by extension) checked UI
boxes.

The currentState object is assigned a value once on initialization
via the Plugin model and Hashmodels, and again on hiberation. The
currentState object is read once when the LayerManager reports that
its resoureces are done loading, and also on activation. This handles
the two uses cases of:
1) A user restores some state from a permalink
2) A user turns the plugin off, and then turns the plugin back on,
revealing the previously selected map layers.

Refactor the UI methods to separate the method that is responsible
for displaying the UI from the method that displays the actual tree
of resources. The render method now explicitly empties its container,
performs some actions on the tree and container, and then displays
itself.