Add region config and forceDeactivate to plugin app object
The plugins should have access to the main app config.  In this case,
the full extent
plugin needed the initial extent for the app.
I added a way for the plugin to deactivate itself rather than wait for
the user to do so.
In this case, full extent has no active state after it moves the map to
it immediately
deactivates iteself, keeping the UI consistent.