Autoselect entire permalink URL
Fix Github Issue #85

Previously the entire permalink url was selected
when the window is created, but there were problems
with users clicking in the box and making a smaller
selection, which was difficult to tell and causing
the URL to not produce the page they desired.

Now, whenever a mouseup event happens in the input
box, the entire URL is selected.

Also make the input box read only.