Allow a clickable buffer around measure origin
In order to make it easier to complete polygons using a touch-screen
device, a buffer area is created around the first node in the polygon
that will trigger polygon completion on click.

The finishMeasureAsPolygon function has been modified to rely on global
state instead of data that is passed in. This was done to make it easier
to call this function from more places.