Permalink to use URL shortener service
The google URL shortener service is free and stable.  When generating
a permalink for app state, it is run through the service.  If it
fails, it will still present the traditional long URL.  The
administrator may optionally provide a googleUrlShortenerApiKey value
in region.json (generated from google API console) which allows
additional requests per day.  No key is required though, and any
usage errors will be logged and the long URL will be used.