Legend can resize, move and re-flow
Reformat the markup generated by the Legend dijit to allow a multi
column reflow when the legend space is enlarged.  Adds support for
resizing and dragging using the existing dojo framework.  I found the
dojo tools to be much weaker than jquery UI, but as this is already a
heavily library-dependent app already, and dojo comes pre-loaded
with the ESRI jsapi, I chose to live with what it can do rather than
add a new dependency.