Fix comparer bug on split screen
Split Screen mode would share and otherwise disrupt the "selected"
graphics from one map to another.  The containers and layer were
initialized when the class was loaded instead of when an instance
was created, so they were actually shared between map panes.  They
are now initialized in the constructor to be unique.