Add support for blacklisting identify attributes
Certain attributes do not provide meaningful data, so this feature allows those attributes to be blacklisted so they can be removed from the output.

As of this commit, you CANNOT blacklist display values, which are very similar to attributes, so it may create the appearance of attributes that cannot be blacklisted.