Remove PageSize and fixup UI markup
PageSize is no longer a parameter for printing, we assume Letter ANSI A
size.  This was to support custom printing, which requires new templates
to be created, so we dropped support for the less used sized.

Also cleaned up the markup so it wouldn't rely so heavily on floats and
paddings to achieve the layout, and closed input tags correctly