Fix eastward migration of split screens
fixes #192 on github.

Use debounce instead of delay to prevent repeated calls. It is clear from my experiments that repeated calls were happening, but not clear if and what aspects of the `esriMap` calls were causing this to miscalculate when recentering multiple times.

I found the problem to persist when the wait param was set to 150, but fixed when bumped up to 300, so probably the browser did not have enough time to finish the resize operation.

Still, debounce is the right behavior here, as little can be gained from repeated calls. In fact, the underscore docs for debounce explicitly mention this use case, "recalculate a layout after the window has stopped being resized."