Fix broken split screen
The Id attribute was used in the ESRI sample, but in addition to
registering a DOM attribute (ID), it also registers and ID in a
weird global registry that doesn't allow duplicates, and was silently
failing.  Use a class instead, which is better markup since there
could be two elements on the screen at once (split screen) and
doesn't have a registry conflict.