Clean up script/css imports
While poking around for the Foundation version (it's 3), I realized the
foundation.js we pull into the project has a version of jQuery and
Modernizr bundled into it.  It appeared that those were the versions
that we ultimately referenced on the page, so I removed the duplicate
declarations.  Presumably, this is not an ideal situation but I think it
should be resolved in a larger script issue (minifiying/concat).