Don't debounce updateLegend
fixes #226 on github

Unfortunately, it's not clear why debouncing causes this not to work, or more precisely, why this doesn't work on the first call but does on the second. However, there will always be two calls when we don't debounce, and my experiments indicated that this fixes the bug, as reported. The answer is hidden in the unavailable source for legendDijit.refresh() (where legendDigit is a esri.dijit.Legend).