Fix layer selector plugins with infographics not rendering.
Since the container node for plugins with infographics is not visible
upon activation, there is no place for ExtJS to render the tree.

To resolve this problem, we added a new method that is called when the
plugin view changes from the infographic to the primary view, and vice
versa.