Fix #265: Fixed Title Bar overlaps portion of top map.
The navbar height was larger than it appeared to be when using the URL
property of the titleDetail setting in region.json. This invisible
bounding box would swallow mouse events and cause dialog windows that
were dragged underneath the title bar to become stuck there.

Foundation CSS has specific line-height and margin settings on the title
bar H1 element but is missing these styles for the H2 element. I copied
these H1 styles for the H2 element to resolve this issue.