Add csv and text export handler
Plugins may generate and display data in their UI, but have needed an
option to be able to export that data.  This handler will echo back an
JSON array of arrays as a csv file, or for more complicated/custom
structures will also simply initiate a download of the posted text.