Implement subregion activation
Subregions can be activated by either clicking on the
subregion shape on the map, or making a selection from the
launchpad "Pick a Geography" section.

Activating a subregion means 1) zooming to the extent of
the subregion (plus a little buffer), 2) turning off the
subregion graphics layers, and 3) triggering an event on
the map to let it know a subregion has been activated.