Make single map launchpad issues act on active map
* If a saved state/launchpad issue contains only one map/pane,
have it update the active map instead of the map the state
was saved for. For example, if the user has the second map
active and they launch an issue that was saved with the first
map active, the second map will now be updated with the saved
state instead of the first map.

Closes #317