Don't show the launchpad on init for permalinks.
If the user came to the page through a URL that has
a hash, i.e. application state, don't show the launchpad
on initalization even if the showByDefault launchpad
config setting is set to true.

The presumption is that if the user was given a permalink,
the person who shared the link wants the user to go right
to the map, and not clear out that saved state by clicking
on an issue on the launchpad.

Closes #333