Use close instead of remove to deactive subregion.
* Calling remove directly instead of calling close didn't
fully clean up and deactivate a subregion when the user
selector another subregion from the launchpad. Calling
remove directly in that context also removed all event
listeners that were assigned on the page, resulting in the
behavior described in #353.
* Also move call to close subregion outside of deactivateSubRegion,
since close calls deactivateSubRegion, and the app would
get stuck in an infinite loop if the function was left in
it's current place.

Closes #353