Remove extra calls to close/deactivate subregion.
* The call to close the current subregion when switching
subregions using the drop-down menu is handled in
activateSubregion and does not need to be called directly
in the drop-down menu event handler.
* Also in activateSubregion, we don't need to make a
direct call to deactivateSubregion since it is made
in subregionHeader.close().
* Having the duplicate calls resulted in the subregion map
border and map element positions to become out of sync
with the status of the subregion. This sometimes meant
in a map border when no subregion was active or map
elements hidden under the subregion header.
* The call to deactivate the subregion when clicking
"free explore" on the launchpad should make sure to also
clear the current subregion id or else the subregion
will not be properly deactivated.