Don't silence change event when updating extent
* This caused the map extent to become stale. It was originally
thought that if the event wasn't silent, the map would get
stuck in an endless loop of saving the extent, firing a change
event, updating the map, then saving the extent, etc. However,
it appears that the ESRI map event 'extent-change' does not
get fired if the extent doesn't actually change.