Always activate plugin when plugin is selected.
In an earlier attempt to decrease the number of events
and function calls related to activating the layer selector
and plugins in general, a check was added to see if a
plugin needed activation before activating it.

However, this did not account for plugins that are minimized
They are active when they are selected by the user, but still
need their activate function called when they are clicked.

Removing this check ensures that a plugin's activation function
is always called when it is selected.

Fixes #412