Comment out deracinate node
Preferred behavior would be to force users to specify in the config that
a node should be placed at the root of the tree as opposed to the
framework automatically doing so; there are some edge cases where the
automatic deracination is undesirable.  Specifying an empty string as
the folder title in the config can emulate the same behavior and makes
it more explicit and transparent for the config developer.