Allow modals to be arbitrary widths
The TINY box popovers were fixed in size via CSS which overrode
the supplied width and height attributes when initializing.  This
removes the override specifically so the print preview window can
be sized, but also updates the existing modals to not supply the
previously ignored size attributes so their apperance is unchanged.