Update jQuery plugins
By moving the order of script loading and updating the scrollbar plugin
(which checks for AMD), we now need to specically load jQuery.  This was
done by dojo before internally.  I matched the version and the source of
the dojo request for jquery and it doesn't attemp to fetch it again.