Add ESRI optimized JS bundle.
This commit adds a minified and bundled JS file containing Dojo, ESRI JS
API, and all of the components of those two libraries that were are
currently using on the app. It was created using the ESRI JS API
Optimizer.

* Use v1.0.4 of the custom ESRI bundle. This includes an updated version
of the ESRI JS API (3.15), and all of the components that are currently
used in the app.
* Replace uses of dojo.require statments, which ignored the bundle file
and made requests to the CDN, with AMD require statements.
* Add a requirements file that can be used to recreate the ESRI JS bundle.
* Make sure ESRI map library is required in the map module, and fix an
issue with the legend's ResizeHandle. The updated version of this
component
doesn't return `this` from the ResizeHandle constructor, so we have to
store the instance first, and then add it to the container.

Refs #448