Attempt to fix legend flickering.
Adding support for tile layer legends (#452) introduced a bug that made
the legend flicker when tile layers were added to the map. This is because
1) tile layers fire multiple events when being added
to the map, which caused our manual layer redrawing to fire multiple
times, and 2) our logic for determining the number of visible layers on
the map did not properly handle tile layers.

Also, the legend was flickering in older version of the framework, most
likely due to changes in the structure of legend elements returned by
ArcGIS server, but it wasn't visible because it was happening
very frequently and quickly.

This commit attempts to fix the legend flickering by reducing the amount
of times the legend is redrawn. It also attempts to make the logic of the
manual legend redrawing hack clearer and easier to follow.

* Try to reduce the number of times updateLegend can be called repeatedly.
* Keep an explicit count of the number of legend nuggets rather than using
the length of the legend nugget array. This was changed because we add
legends for layers on the same service in groups now rather than
breaking them up. This preserves the legend hierarchy and is a littler
cleaner and less disruptive to the original structure
* Rely on a map method and direct property of the layers for determining
which layers are visible, so that the logic is more clear.
* Reduce the delay from when update legend is called again if the visible
layer count doesn't match. Testing showed that the layers were restored
* more quickly than 1 second.

Refs #471