Reposition legend if it's cut-off by viewport.
* It's possible that the legend will be partially off screen
when it's auto-resized. This happens if layers were added when
the legend was minimized, or in some cases, if the user moves
the legend. After resizing the legend, check to make sure it's
not cut off by the right or bottom side of the viewport. If it is,
reposition it so that it's completely visible.

Refs #509