Check custom legends to determine legend redraw
This is a stopgap measure to prevent legend flickering when using the
Habitat Explorer.  The comparison of layers on the map vs legends in the
Legend is expanded to account for possible custom legend.  The need to do
the comparison is a result of the existing unreliability of the JSAPI to
notify of `layerAdd` events.