Hide "No Legend" message
Occasionally, the ESRI Legend digit will print a message "No Legend" when
there are no legends (this seems to only happen in split screen mode).
Since we have custom legends, the digit doesn't know that there are
legends, so the message is incorrect.  This hides the message since we
aren't investing in the ESRI Legend component in the future.