Track legend non-participating services
In the list of layers that should not be counted to have a legend region,
include non-feature layer services which have not opted out of the
framework legend.  This prevents mis-counting in regions that use the
Community Rating System plugin, which adds layers to the map but requests
that they not be included in the legend.