Remove getServiceIdForLayer
In the places where this was used, we already have access to the
service layer, which makes this redundant.

Also, this function is nearly identical to findLayer, which I didn't realize
during code review. I prefer to use findLayer instead since it is more
generic and handles an edge case that is missing in
getServiceIdForLayer.