Update Gulf of Mexico Florida layers
This fixes a dojo error that occurs when the page loads. The cause is
due to an ID mismatch between layers.json and the map service. Looks
like the map service was updated recently, so this change should bring
the layers.json file back in sync with the map service.