All infoGraphic to be an HTML snippit or an img...
The snippit allows the code to detect if the infoGraphic is specified as an image (.png or .jpg).  If not it assumes that it is a an html snippit and inserts into the same location.  Done at the request of Zach, should not impact current plugins that specify an image.