Redraw legend upon map resize
Allow plugins to redraw the legend by calling `map.resize()`.

This is a workaround for the fact that adding custom legend HTML does
not trigger any events we listen to for redrawing the legend.

If custom legend HTML is created while the legend is minimized, it will
stay minimized.

I decided that it would be easier for plugins to trigger an event on the
`map` instance (that each plugin already has access to) instead of
updating the API to pass in the current `Legend` instance.

This was previously not an issue because the legend container remained
open at all times, even when it was empty. Plugins that rely on this old
legend behavior, and use custom legend content, will need to be updated
to call `map.resize()` each time custom legend content is updated.

Connects #665