Allow features to be identified based on Geometry
To account for small and large feature types (polygon vs else), two
identify requests are made at different tolerances.  For lines, the code
assumed there is a `Shape` attribute to make the decision about which
group of feature a particular result belonged to.  This was only the case
for geodatabase types that include that attribute automatically, and in
some cases it didn't exist. By including the geometry type in the returned
results, this check can be made consistent.