Improve visibility of plugin error message
There is a lot of noise in the console when there are many plugins being
used with the framework. This change makes it a little easier to see when
there was a problem initializing a plugin.

Also, the message logged from e is present when logging e.stack. To reduce
some of the noise, the log of e is removed here.

Refs #660