Integrate new plugin design
Add plugin template from new design and remove unneeded style rules.

Update selectors in plugin code to reference new template.

Remove calls to size setting functions. This will be handled in #696.

Implement plugin icon property. This replaces the plugin image.

Remove references to second pane in app initialization, which was causing
problems in plugin initialization. I also reverted the change to the first
pane selector, so that it was an id rather than a class.

Refs #686