Allow plugins to deactivate themselves
Plugins are traditionally activated and deactivated by clicking on the
plugin launcher button or the plugin window toolbar buttons. However,
it is sometimes desirable for a plugin to have the ability to deactivate
itself.

The map_utils plugin is an example of this. For the measure command, we
want the plugin to remain active until deactivated by the user. For other
commands such as "zoom to full extent" we want the plugin to deactivate
immediately after executing the command.