Implement custom color overrides in new design
The CSS rules in the index document that use the custom colors were
updated with selectors that are used in the new design.

Also, a third color was added to the region config. This color is used
to configure the color of the sidebar.

The default colors were updated to match the new design.

Refs #695