Fix measure plugin tooltips
Style rules for the new design made the text for the tooltip white on
white. We explicitly set it to the font color used in other parts of
the app so that it is legible.

Also, a certain plugin adds the "claro" Dojo/Dijit theme to the body
element. This adds an undesirable background image to the measure
tooltip, which we override and dispose of here.