Setup basic map print workflow
Adds and wires-up a basic implementation of the new map printing workflow.
This workflow mimics the plugin print workflow. Code related to the AGS
print service is removed because we will be relying on print stylesheets
instead.

A new region configuration was added that allows site admins to specify
the logo to be used on the printed map page.

Refs #724