Add new launchpad
Adds the new, redesigned launchpad that functions more like a plugin.

The new launchpad supports a list plugins, saved states, and sponsors (to
be implemented in #741). Subregions are no longer supported. To support
these changes, the region.json schema and sample file are updated here.
Otherwise, the existing launchpad schema is repurposed.

Plugins can be launched from the launchpad, so the plugin API is modified
to allow the plugin to be activated via an event from the launchpad.

The general plugin architecture follows the design of the regional-planning plugin.

Refs #736