Expect plugin sidebar icon to be named icon_sm.png
To avoid having to maintain a `prototype` branch for every plugin, we
now use icon_sm.png instead of icon.png for the plugin sidebar icon.

This will allow plugins to use the old icons when running in the
context of the `develop` branch, and use the new icons in the context of
the `prototype` branch.

Connects #743