Add API to support in app help
Plugins can register with the framework to acknowledge that there is
in-app help functionailty available.  The framework handles the UI of
showing the help icon in the header and for managing the state of "show on
start" via local storage.  Its intended purpose is to allow plugins to
manage their own help functionality and move away from the exising
infographic approach.