Make layer info box visible
After the markup was changed to implement the new style, seeing
the layer info box required scrolling to the right. This is because
of the overflow-y style on parent nodes. Removing this style
makes it so the info box appears to the side of the layer selector
without needing to scroll.