Add back overflow styling to sidebar
A previous PR https://github.com/CoastalResilienceNetwork/GeositeFramework/pull/838
changed the overflow styling to make the Regional Planning info box visible.
Unfortunately, it broke the scrolling on the Launchpad, so we undo that PR
in this commit. A new commit to the Regional Planning plugin repo solves
the problem of making the info box visible in a different way.