Remove extra jQuery module, update for Bootstrap
When the new design was implemented, a new version of jQuery was added
to the framework to support Bootstrap. Aan existing version of
jQuery, too old to support Bootstrap, was already loaded on the page.

However, the existing jQuery module just needed a minor version update
to support Bootstrap. That change is made here, and the newer version of
jQuery is dropped.