Add tool to expand and shrink map
Add button to the map that allows the user to expand and shrink the size
of the map by collapsing the sidebar.

This is also done automatically at small screen sizes when plugins are
activated, so the tool listens for changes to the map size so that the
icon can be appropriately updated (or else it would become out of sync).

Since the CSS rules for the collapsed sidebar were in a media query,
they are copied and placed outside of the media query as well so that
they can be used at all screen sizes. This is not ideal.

Also, since the sidebar can now be minimized at any screen resolution,
the rule to hide the close button when the sidebar is minimized was
moved outside of a media query so that it would apply at all times.

Refs #887