Fix and improve the map expand/shrink
The initial implementation of this tool duplicated the CSS rules that
were contained in a media query to make the sidebar narrow. Doing this
meant that the sidebar became narrow at any screensize once a plugin was
activated. To fix this, a different class name is now used for the tool.

Also, the logic to check whether or not the sidebar is narrow was improved
here to take the size of the viewport into account.

To better keep the tool button icon in sync with the state of the sidebar,
changes to the viewport size are listened to and the icon is updated if
need be.

One side-effect of having a tool to manually toggle the state of the
sidebar is that users may have to expand the sidebar using the button
on small screens if the sidebar was initially toggled manually and any
active plugins are turned off.

Refs #912