Fix app-print cleanup
Currently the app-print.css file, with class 'app-print-css' is not getting removed from the app when closing the print dialog because the class is created with a "." in it.  The class shouldn't be prepended with a "." in the class attribute.  Removing fixes the problem and the css file is correctly removed on cleanup. c.c. @zferdana