Remove legacy sidebar code
The code and templates removed here should have been removed when the
new design was implemented. They are removed here because the showExportButton
implementation is no longer used, but may be confusing since this commit will
be followed up with a different implementation of hiding and showing the
export tool.

Refs #979