Use original map in plugin print workflow
Instead of creating a new map when printing from a plugin, reuse the
existing main map. This will make it easier for plugin developers to
print maps that match the main map, which appears to be the main use
case. Code was added to move the map around from the main area, to the
plugin print sandbox, and then back to the main area after printing is
complete.

The css was adjusted to not hide the main map, as well as provide a
minimal amount of default styling to the print-out. The rules are can be
overidden and should provide a good starting point for plugin developers
to style their own layouts.

Refs #998