Add support for A4 page size for printing
Adds stylesheets for A4 page size (both landscape and portrait
orientations) as well as an option to the framework print dialog that
allows users to select their page size. The A4 stylesheets were based
of the letter size stylesheets, but with difference in dimensions
between A4 and letter added/subtracted.

The existing page orientation stylesheets were renamed to reflect that
they are for letter size.

Plugins can utilize the page size and orientation stylesheets either by
adding them to the page directly when printing, or getting user feedback
in the plugin print dialog (to be added in #1009) and adding the
appropriate stylesheets.

Refs #985