Clean-up identify point plugin
Cleans up the UI and the internals of the identify point plugin.

The sample form controls are moved to a secondary tab, inline style
rules are eliminated or moved to an external CSS file, and a template is
used instead of jQuery to build out the UI.

Also, a print button is added to the UI in single plugin mode. This
button will be wired up in a subsequent commit.

Refs #1006