Assign events after plugin content is rendered
Moves code around so that UI events are assigned after the content for a
plugin is in the DOM. This allows plugins, particularly when in single
plugin mode where the plugin header is hidden, to add elements that can
be hooked into the standard plugin events.

This is most useful for the printing, where plugins will be adding a
custom print button in single plugin mode that needs to activate the
built-in plugin print workflow.

Refs #1006