Implement infographic support in the launchpad
Add support for the infographic feature that was developed by TNC plugin
authors.

Site admins can now enable this feature in the region config, provide a
HTML template, and a button will be added to the launchpad plugin title.
When users click this button, a modal infographic is displayed.

This implementation is largely based on the feature implemented in
https://github.com/CoastalResilienceNetwork/ecological-effects-slc.

Refs #939