Include plugin legends in printed map legend
Previously, plugin legends were not displayed in the printed map legend
because they were not accounted for in the print stylesheets. This resulted
in the legends being rendered off the page. Now, the plugin legends container
is temporarily moved into the layer legends container. That, combined with
modifications to the print stylesheets, makes the plugin legends visible
on the printed page.

In the course of this work, the general legend print styling was also
improved and a superfluous wrapping of legend elements in a special
print class was removed.

It should be noted that while this does allow for the plugin legends to
be displayed in the printed legends, most plugin legends are SVG and
will not display well. They will likely be taller than the legend area,
which will result in the legends being cut-off. However, the scope of
this PR was just to get them to display. Plugin authors will be
responsible for getting the legends to display well.

Refs #1001