Implement infographic for plugins
Adds a configuration option to plugins for infographics. When enabled,
a button is added to the plugin title bar, which, when clicked, will
open the provided infographic content in a modal.

Note that this approach is similar to what was implemented for the
launchpad in #1020. However, because of differences in configuration
(the launchpad is configured via the region.json file, while the plugins
are configured directly in the code), the implementations could not be
merged.

Refs #1025