Transform print preview map into print modal
Repurpose the print preview map into a customizable print modal.
Plugin authors can provide content, like a form, that will be used
to populate the modal. Then, before printing is kicked-off, the
plugin will again be given a reference to the modal content. A
plugin author can parse the content for user input, and then adjust
the print output as necessary.

Refs #1009