Ensure underscore is available through Use config
Import underscore locally and add it to the Use config so that plugins
can reference it.

Plugins that don't directly import underscore reference it through
Use, but it is properly made available through that interface.
Because of that, require.js tries to find it in the ESRI/DOJO package
and can't, which causes underscore to be undefined.

Refs #1043, #669