Fix layer alignment in framework print
Resize and reposition the map each time it is moved in the DOM or
resized manually. This ensures that layers are redrawn in the correct
location when the map is refreshed, and that the map generally behaves
normally when it is interacted with after printing.

The helper function to do this is named after the Leaflet function that
performs a similar purpose.

Refs #1035