Hide infographic button in single plugin mode
This button was designed for the plugin header, which doesn't exist in
single plugin mode. Also, the "About" pane acts as a replacement for the
infographic.

The function to add the infographic button is moved out of the render
function and into to the plugin initialization function. Placing it in
the render function resulted in multiple buttons being added since the
render function is designed to be called multiple times.