Fix rendering of plugin legends in print
Changes introduced in #1051 exposed an issue with the solution
introduced in #1027 to include plugin legends on printed maps.

In #1027, the plugin legends DOM element was moved into the layer
legends DOM element for printing. However, the layer legends DOM
element is dynamically updated and gets redrawn when the map updates.
Because the changes in #1027 force many more map updates to
ensure layers don't get out of sync when the map size changes, the
plugin layers element is overwritten when the layer legends element is
rerendered.

I was able to remove the solution introduced in #1027, and make a much
simpler fix, which is to move the plugin legends element ahead of the layer
legends element in the DOM. This does have the side-effect of placing the
plugin legends first in the legend container. It's not totally clear why this
solution works, other than it being a result of the CSS rules in place
for the legend during printing. Much work has already been invested in
these rules, so I did not want to modify them any further. I did, however,
include some new style rules for plugin legends to add a minimal set of
constraints on them since they are often very different from one another.

Refs #1050