Attempt to support various underscore.js imports
Adjust the use config to support all of the various ways that underscore
is imported within plugins.

The previous fix implemented in #1044 caused problems with plugins that
bring in their own version of underscore in via a CDN. The adjustments
here should make that possible while still allowing plugins to import
underscore via Use.js.

Also, rename the framework version of underscore to reduce confusion.

Refs #669, #1043, #1046