Unify prePrintModal and postPrintModal arguments
Adjusts the arguments for prePrintModal and postPrintModal so that all
of the arguments provided to one match the other, and all of the
arguments have the same name and are in the same order. The print
workflow example in the identify point plugin is adjusted to take these
changes into account.

Refs #1057